feat: add all functionality from the fourth lesson
