+++ src/components/ChatList.js
@@ -1,9 +1,12 @@
+import { Link } from "react-router-dom"
+
 const ChatList = ({ chats, chatId }) => {
+
     return <div className={'chatList'}>
         {Object.keys(chats).map((chat, i) => (
-            <div key={i}>{chats[chat].name} </div>
+            <Link to={`/chats/${chat}`} style={{ color: chat === chatId ? 'grey' : 'black' }} key={i}>{chats[chat].name} </Link>
         ))}
-    </div>
+    </div >
 }
 
 export default ChatList
