+++ src/components/MessageList.js
@@ -1,15 +1,44 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { List, ListItem, ListItemText, Box, ListItemAvatar, Avatar } from '@mui/material'
+import { Adb, Face } from '@mui/icons-material';
+
 
 
 const MessageList = ({ messages }) => {
-    return <div className='messages_space'>
-        {messages.map((message, id) => (
-            <div className={`messages_${message.author === 'human' ? 'human' : 'robot'}`} key={id}>
-                {message.text} - say: {message.author}
-            </div>
-        ))}
-    </div>
+    return (
+        <Box
+            sx={{
+                width: 300,
+                height: 600,
+                border: '1px solid #ccc',
+                overflow: 'auto'
+            }}>
+            <List sx={{ mb: 2 }}>
+                {messages.map((message, id) => (
+                    <ListItem
+                        button
+                        key={id}
+                    >
+                        <ListItemAvatar>
+                            <Avatar sx={{ bgcolor: message.author === 'bot' ? 'green' : '#1d46ad' }}>
+                                {message.author === 'bot' ? <Adb /> : <Face />}
+                            </Avatar>
+                        </ListItemAvatar>
+                        <ListItemText primary={message.text} secondary={message.author} />
+                    </ListItem>
+                ))}
+            </List>
+            {/* <div className='messages_space'>
+                {messages.map((message, id) => (
+                    <div className={`messages_${message.author === 'human' ? 'human' : 'robot'}`} key={id}>
+                        {message.text} - say: {message.author}
+                    </div>
+                ))}
+            </div> */}
+        </Box>
+
+    )
 }
 
 MessageList.propTypes = {
